requirements.txt
cryptography==43.0.1
fastapi[standard]
huggingface_hub==0.25.1
langchain_community==0.3.2
langchain_experimental==0.3.2
numpy
pgvector==0.1.8
psycopg[binary]
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
sentence_transformers==2.2.2
tiktoken==0.7.0
unique_names_generator==1.0.2
uvicorn
gunicorn
supabase
openai
python-multipart